% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recruitment.R
\name{BH.rec}
\alias{BH.rec}
\title{Beverton Equation}
\usage{
BH.rec(sp, bha, bhb, bio, ver = "N")
}
\arguments{
\item{sp}{spawning power}

\item{bha}{Alpha parameter}

\item{bhb}{Beta parameter}

\item{bio}{Biomass}
}
\value{
The amout of recruit (Larvaes)
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Demiurgo
}
