% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pprey.mat.R
\name{Bio.ages}
\alias{Bio.ages}
\title{Biomass by age class}
\usage{
Bio.ages(Biom.N, age, Over.mat)
}
\arguments{
\item{Biom.N}{Biomass by Cohort}

\item{age}{Age of maturation}

\item{Over.mat}{Overlap matrix}
}
\value{
Biomass by adult and juveniles
}
\description{
Biomass by age class
}
\author{
Demiurgo
}
