% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pprey.mat.R
\name{Bio.func}
\alias{Bio.func}
\title{Biomass from nc file}
\usage{
Bio.func(nc.file, groups.csv, numlayers)
}
\arguments{
\item{nc.file}{Atlantis initial condition file}

\item{groups.csv}{Atlantis groups file}

\item{numlayers}{Number of layers by box}
}
\value{
The biomass for age class and the sturctural nitrogen by age class
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Demiurgo
}

