% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{bio.pwn}
\alias{bio.pwn}
\title{Biomass for age groups}
\usage{
bio.pwn(pwn.grp, nc.out, ctg, mg2t, x.cn, box.info, Time)
}
\arguments{
\item{pwn.grp}{Biomass pools with age classes}

\item{nc.out}{ncdf atlantis' output file}

\item{ctg}{category}

\item{mg2t}{mg C converted to wet weight in tonnes == 20 / 1000000000}

\item{x.cn}{Redfield ratio of C:N 5.7}

\item{box.info}{Information for each box and layer}

\item{Time}{Dates vector}
}
\value{
a dataframe with the biomass for all the functional groups with Biomass pool age classes
}
\description{
Biomass for age groups
}
\author{
Demiurgo
}
