% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{boxes.prop}
\alias{boxes.prop}
\title{Box information}
\usage{
boxes.prop(bgm.file, depths)
}
\arguments{
\item{bgm.file}{BGM file,  Atlantis input}

\item{depths}{Cummulative depths (de max depth of each layer)}
}
\value{
a dataframe with information (box-id, area (m), )
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Demiurgo
}

