% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catch.R
\name{catch}
\alias{catch}
\title{Analysis of the Harvest in Atlatnis}
\usage{
catch(grp.csv, fish.csv, catch.nc, ext.catch.f = NULL)
}
\arguments{
\item{grp.csv}{Groups csv file (Atlatnis input file)}

\item{fish.csv}{Fisheries csv file (Atlantis input file)}

\item{catch.nc}{Netcdf Output file from Atlantis. Usually with the name of [Your_Model]CATCH.nc}

\item{ext.catch.f}{External File with the Observed catches and discards by year. This helps to calibrate the harvest section of atlatnis.}
}
\value{
A shiny output (reactive html)
}
\author{
Demiurgo
}
