% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare}
\alias{compare}
\title{Compare two diferent outputs}
\usage{
compare(nc.out.current, nc.out.old = NULL, grp.csv, bgm.file, cum.depths)
}
\arguments{
\item{nc.out.current}{Current netcdf output file. Character string with the connection to the netcdf file to read in. This netcdf file contains is a generic output from an Atlantis run usually starts with \code{output} and ends in \code{.nc}".}

\item{nc.out.old}{(Default = NULL) This is the old netcdf output file,  used only for comparation. Character string with the connection to the netcdf file to read in. This netcdf file contains is a generic output from an Atlantis run usually starts with \code{output} and ends in \code{.nc}".}

\item{grp.csv}{Character string with the connection to the Groups \code{*.csv} file (Atlantis input file).}

\item{bgm.file}{Character string with the connection to the XY coordinates Antlantis input file \code{*.bgm} with the information in meters.}

\item{cum.depths}{Cummulative depth of the Atlatnis model}
}
\value{
A shiny output (reactive html)
}
\author{
Demiurgo
}
