% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality.R
\name{.calc.mort}
\alias{.calc.mort}
\title{Extranting mortality values from the txt files provided by Atlantis}
\usage{
.calc.mort(SpeMort, PredMort, FG)
}
\arguments{
\item{SpeMort}{Especific mortality}

\item{PredMort}{Predation mortality}

\item{FG}{functional group}
}
\value{
List of mortalities by functional group by time
}
\description{
Extranting mortality values from the txt files provided by Atlantis
}
\author{
Javier Porobic
}
