% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{.point_in_polygon}
\alias{.point_in_polygon}
\title{Raycasting Algorithm to find out whether a point is in a given polygon.
Performs the even-odd-rule Algorithm to find out whether a point is in a given polygon.
This runs in O(n) where n is the number of edges of the polygon.}
\usage{
.point_in_polygon(polygon, point)
}
\arguments{
\item{polygon}{an array representation of the polygon where polygon[i,1] is the x Value of the i-th point and polygon[i,2] is the y Value.}

\item{point}{an array representation of the point where point[1] is its x Value and point[2] is its y Value}
}
\value{
whether the point is in the polygon (not on the edge, just turn < into <= and > into >= for that)
}
\description{
Raycasting Algorithm to find out whether a point is in a given polygon.
Performs the even-odd-rule Algorithm to find out whether a point is in a given polygon.
This runs in O(n) where n is the number of edges of the polygon.
}
\author{
Javier Porobic
}
