% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pprey.mat.R
\name{feeding.mat}
\alias{feeding.mat}
\title{Atlantis feeding tool}
\usage{
feeding.mat(prm.file, grp.file, nc.file, bgm.file, cum.depths,
  quiet = TRUE)
}
\arguments{
\item{prm.file}{Character string with the path to the biological parameter \emph{*.prm} file (Atlantis input file).}

\item{grp.file}{Character string with the path to the Groups \emph{*.csv} file (Atlantis input file).}

\item{nc.file}{Character string with the path to the netcdf file to read
in. This netcdf file contains the initial conditions for the Atlantis model
usually ends in \emph{.nc}.}

\item{bgm.file}{Character string with the path to the XY coordinates Atlantis input file \emph{*.bgm} with the information in metres.}

\item{cum.depths}{Vector with the cumulative depths of the different layers
(e.g. \emph{cum.depths <- c(0, 20, 100, 200, 500)})}

\item{quiet}{(Default = TRUE) this parameter helps during the process of debugging.}
}
\value{
Shiny::Reactive shiny::HTML that displays predator-prey information such as: the ppPREY
    matrix, the initial abundance of prey, the overlap matrix based on gape size,
    predator preference and predator prey spatial overlap. All this information
    is divided into two different tabs:
 \itemize{
  \item \bold{Non-spatial}: This tab allows the user to check and modify
    availability values from the pprey matrix. This Function provides 5 outputs:
   \itemize{
     \item \bold{Availability matrix (\eqn{\lambda})}: Matrix of prey
    availability values (pprey) for each adult, young or biomass pool prey and
    predator.
     \item \bold{Overlap matrix}: Calculates if a predator is able to eat a prey
    given its gape size limitations. This function uses the knife-edge size
    selectivity, where availability of the prey is either available (1) or not
    available to the predator (0).
 \deqn{\omega = \left\{
    \begin{array}{ll}
    1 & SN_{predator} * KLP < SN_{prey} < SN_{predator} * KUP \\
    0 & Otherwise
    \end{array}
    \right. }{ (non-Latex version) }
 were \eqn{SN} is the structural weight; \eqn{KLP} Minimum gape limit of the
    predator (age structured or biomass pool); \eqn{KUP} Maximum gape limit of
    the predator (age structured or biomass pool).
\item \bold{Effective Predation}: this provides an approximation of the total
    biomass of prey consumed by a predator. It assumes a perfect spatial match
    between prey and predator.
    \deqn{\gamma  =  B * \lambda * \omega}
were \eqn{\gamma} is the effective predation; \eqn{B} is the biomass;
    \eqn{\lambda} is the predator-prey availability (pprey matrix); and
    \eqn{\omega} is the predator-prey gape overlap.
\item \bold{Predation pressure}: this function calculates the percentage of the
    predator diet corresponding to a specific prey.
\item \bold{Total biomass prey}: this is the total biomass per functional group
    and maturity stage (adult and juvenile).
}
\item \bold{Spatial Overlap}- this tab allows the user to check the spatial
    overlap between the prey and the predator in all the boxes and layers based
    on the values in the initial conditions file and the parameterized gape
    limitation.
}
}
\description{
This function helps the user to calibrate and explore different aspect of the
    predator-prey interaction. In Atlantis, interaction between the predator and
    the prey is mainly defined by the predator-prey matrix which represents the
    maximum availability of each prey (age) group to a specific predator (age
    group). This consumption can be shiny::strongly affected by different processes such
    as the spatial overlap, the biomass of the prey and the gape limitation of
    the predator. With this tool you can explore the availability matrix and the
    processes that affect consumption. This tool also allows you to explore new
    values for the predator-prey matrix and shiny::observe online how these new values
    affect the predator-prey interaction.
}
\author{
Demiurgo
}
