% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pprey.mat.R
\name{find.z}
\alias{find.z}
\title{Num layer  per box}
\usage{
find.z(bgm.file, cum.depths)
}
\arguments{
\item{bgm.file}{BGM file for the atlantis model}

\item{cum.depths}{Cummulativce depths of the model}
}
\value{
dataframe with the number of layer per box including the box that are island
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Demiurgo
}

