% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pprey.mat.R
\name{gape.func}
\alias{gape.func}
\title{Calculate the gape limitation for each functional group}
\usage{
gape.func(groups.csv, Struct, Biom.N, prm)
}
\arguments{
\item{groups.csv}{Atlantis group file}

\item{Struct}{Structural weight by age group}

\item{Biom.N}{Biomass by age groups}

\item{prm}{Atlantis paramter file}
}
\value{
the limits for the prey based on the predator gape size and prey size
}
\author{
Demiurgo
}

