% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth.R
\name{growth.pp}
\alias{growth.pp}
\title{Growth primary producers}
\usage{
growth.pp(ini.nc.file, grp.file, prm.file, out.nc.file)
}
\arguments{
\item{ini.nc.file}{Atlantis initital condition file (netcdf file)}

\item{grp.file}{Group file. Atlantis input file (.cvs)}

\item{prm.file}{Atlantis Biology parameter file}

\item{out.nc.file}{Estandar netcdf output file}
}
\value{
A shiny output (reactive html)
}
\author{
Demiurgo
}
