% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pprey.mat.R
\name{make.map}
\alias{make.map}
\title{Function to get the vertices of the bgm map for Atlantis}
\usage{
make.map(bgm.file)
}
\arguments{
\item{bgm.file}{BGM file for atlantis}
}
\value{
A dataframe con witht 3 columns,  first the box_id,  then the latitude and the longitude
}
\author{
Demiurgo,  based on Shane function \href{https://github.com/shanearichards/shinyrAtlantis}{shinyrAtlantis}
}
