% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality.R
\name{mortality}
\alias{mortality}
\title{Mortality analysis for Atlantis}
\usage{
mortality(grp.file, prm.file, SpeMort, PredMort)
}
\arguments{
\item{grp.file}{Character string with the path to the groups \emph{*.csv} file (Atlantis input file).}

\item{prm.file}{Parameter file}

\item{SpeMort}{Character string with the path to the plain text Specific
mortality file \emph{*SpecificMort.txt} (Atlantis output file).}

\item{PredMort}{Character string with the path to the plain text predation
morality file \emph{*SpecificPredMort.txt} (Atlantis output file).}
}
\value{
A reactive output with plots to analyse mortality in Atlantis
}
\description{
This function helps to explore the different types of mortalities in an Atlantis
    model. Technically, this function takes the results of mortality from
    Atlantis and compares them. This includes temporal components of natural, fishing and predation
    mortality. This function also explores the different mortalities in space and
    time.
}
\author{
Javier Porobic
}
