% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{nitro.weight}
\alias{nitro.weight}
\title{Calculate the different weigt (estructural reserve) and number of indivduals}
\usage{
nitro.weight(nc.out, grp, FG, By = "Total", box.info, mg2t, x.cn)
}
\arguments{
\item{nc.out}{Netcdf out file. this is the traditional .nc file from atlantis}

\item{grp}{groups csv file}

\item{FG}{Functional group (selected)}

\item{By}{option to aggregate the time series in just one or leave the values by cohort}

\item{box.info}{Internal function}
}
\value{
List witht he weight (reserve and structural), total biomass (tons) and number
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Demiurgo
}

