% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/food.web.R
\name{ord}
\alias{ord}
\title{Piramid order}
\usage{
ord(vector)
}
\arguments{
\item{vector}{Vector of numbers or character that needs to be order with the bigest value in the center}
}
\value{
A vector with the higest value in the center
}
\description{
Piramid order
}
\author{
Demiurgo
}
