% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{output.cal}
\alias{output.cal}
\title{Compare two diferent outputs}
\usage{
output.cal(b.old.file, b.curr.file, groups.csv, diet.file, age.biomass = NULL)
}
\arguments{
\item{b.old.file}{Old output biomass file from Atlantis.  usually 'output[YOUR_Model_NAME]BiomIndx.txt'}

\item{b.curr.file}{Current output biomass file from Atlantis.  usually 'output[YOUR_Model_NAME]BiomIndx.txt'}

\item{groups.csv}{group csv file}

\item{diet.file}{output file for the diet (atlantis output)}

\item{age.biomass}{output file for the biomass by age (non-standard atlantis output, make sure to put the flag 'flag_age_output' in the run.prm if you want to look at this output)}
}
\value{
A shiny output (reactive html)
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Demiurgo
}

