% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pprey.mat.R
\name{Over.mat.func}
\alias{Over.mat.func}
\title{Overlap matrix}
\usage{
Over.mat.func(Ava.mat, Gape)
}
\arguments{
\item{Ava.mat}{Availavility matrix (or pPREY matrix from the parameter file}

\item{Gape}{Limit of prey by Gape size}
}
\value{
Overlap Matrix based only in the gape limitation
}
\author{
Demiurgo
}

