% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{plot.age.total}
\alias{plot.age.total}
\title{Interactive Plot for biomass (structural and reserve nitrogen)}
\usage{
\method{plot}{age.total}(total, Time, rn2, sn2, num2, bio2, scl2, limit, right,
  colors, coh = NULL, max.coh = NULL)
}
\arguments{
\item{total}{Dataframe with the weight output from the nitro.weight function}

\item{rn2}{True if active the plot of reserve nitrogen}

\item{sn2}{True if active the plot of structural nitrogen}

\item{num2}{True if active the plot of numbers}

\item{bio2}{True if active the plot of biomass}

\item{scl2}{True if active the scaling base on the initial value of the time serie}

\item{limit}{True if active limiting the plot area from 0 to maximum 3}

\item{right}{choose left of right for the legend of the plor}
}
\value{
A interactive plot
}
\author{
Demiurgo
}

