% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catch.R
\name{plot.catch}
\alias{plot.catch}
\title{Plot catch function}
\usage{
\method{plot}{catch}(
  ctch,
  Time,
  ylm = NULL,
  coh = NULL,
  col.bi,
  bio.n = NULL,
  by.year = FALSE,
  external = NULL,
  ...
)
}
\arguments{
\item{ctch}{Catch data.frame}

\item{Time}{Time on Date format}

\item{ylm}{Ylimit option for comparison}

\item{coh}{cohort number}

\item{col.bi}{Color of the lines}

\item{bio.n}{Name of the FG}

\item{by.year}{Plot and information is by year}

\item{external}{External file}

\item{...}{Other options for plotting}
}
\value{
plot of the catch
}
\description{
Plot catch function
}
\author{
Demiurgo
}
