% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/food.web.R
\name{plot.tlvl}
\alias{plot.tlvl}
\title{Plot trophic levels}
\usage{
\method{plot}{tlvl}(T.lvl, rel.prey, foc.fg, pol = NULL, color.p)
}
\arguments{
\item{T.lvl}{Trophic level by predator}

\item{rel.prey}{Realized prey}

\item{foc.fg}{Focus funciontal group}

\item{pol}{Polygon}

\item{color.p}{Colors used}
}
\value{
A food web plot
}
\description{
Plot trophic levels
}
\author{
Javier Porobic
}
