% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recruitment.R
\name{recruitment.cal}
\alias{recruitment.cal}
\title{Estimation of Recruitment function}
\usage{
recruitment.cal(ini.nc.file, out.nc.file, yoy.file, grp.file, prm.file,
  quiet = TRUE)
}
\arguments{
\item{ini.nc.file}{Atlantis initital condition file (netcdf file)}

\item{out.nc.file}{Estandar netcdf output file}

\item{yoy.file}{Young of the Year output file the estandar name for this file is : output[YOUR_MODLE]YOY.txt where [YOUR_MODLE] is the name of yout atlatnis model}

\item{grp.file}{Groups csv file (Atlatnis input file)}

\item{prm.file}{Biology parameter file (*.prm file)}

\item{quiet}{(Default = TRUE) this paramter helps the debuging}
}
\value{
Reactive output whith recruitment curves and estimated values
}
\author{
Demiurgo
}

