% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recruitment.R
\name{Ricker.rec}
\alias{Ricker.rec}
\title{Ricker recruitment equation}
\usage{
Ricker.rec(ralpha, rbeta, biomass, std = TRUE)
}
\arguments{
\item{ralpha}{alpha parameter from the ricker recruitment equation}

\item{rbeta}{beta parameter from the ricker recruitment equation}

\item{biomass}{Stock biomass}

\item{std}{(bolean) if TRUE ricker equation or FALSE Baltic sea version}
}
\value{
The recuitment prooduced
}
\description{
Ricker recruitment equation
}
\author{
Javier Porobic
}
