% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pprey.mat.R
\name{saveData}
\alias{saveData}
\title{Save pPREY matrix}
\usage{
saveData(data)
}
\arguments{
\item{data}{matrix of pPrey modified}
}
\value{
a txt file of the pPREY matrix
}
\description{
Save pPREY matrix
}
\author{
Demiurgo
}
