% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pprey.mat.R
\name{sepText}
\alias{sepText}
\title{text separatror}
\usage{
sepText(ortext)
}
\arguments{
\item{text}{a string scalar}
}
\value{
3 columns with names and stage of the predator and the prey
}
\author{
Demiurgo
}
