% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catch.R
\name{skill_assessment}
\alias{skill_assessment}
\title{Skill assessment of the model}
\usage{
skill_assessment(obs, mod, FG)
}
\arguments{
\item{obs}{Shiny::Observed values}

\item{mod}{modeled values}

\item{FG}{Functional group}
}
\value{
metrics  =  AAE; AE; MEF; RMSE; COR
}
\description{
Skill assessment of the model
}
\author{
Demiurgo
}
