% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Util-text.R
\name{text2num}
\alias{text2num}
\title{Parameter file reader}
\usage{
text2num(
  text,
  pattern,
  FG = NULL,
  Vector = FALSE,
  pprey = FALSE,
  lineal = FALSE
)
}
\arguments{
\item{text}{Biological parametar file for Atlatnis}

\item{pattern}{Text that you are looking}

\item{FG}{Name of the functional groups}

\item{Vector}{Logic argument, if the data is on vectors or not}

\item{pprey}{Logic argument, if the data is a pprey matrix or not}

\item{lineal}{Logic argument,  if the data is in a vector a lineal}
}
\value{
A matrix with the values from the .prm file
}
\description{
Parameter file reader
}
\author{
Javier Porobic
}
