% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catch.R, R/theme_atlantis.R
\name{theme_atlantis}
\alias{theme_atlantis}
\title{Main Atlatnis Theme}
\usage{
theme_atlantis()

theme_atlantis()
}
\value{
The ggplot theme

The ggplot theme
}
\description{
Main Atlatnis Theme

Main Atlatnis Theme
}
\author{
Javier Porobic

Javier Porobic
}
