% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{time.calc}
\alias{time.calc}
\title{NCtime calculation}
\usage{
\method{time}{calc}(ncfile = NULL)
}
\arguments{
\item{ncfile}{Netcdf file}
}
\value{
Vector of dates
}
\description{
NCtime calculation
}
\author{
Javier Porobic
}
