% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Util-text.R
\name{time_calc}
\alias{time_calc}
\title{NCtime calculation}
\usage{
time_calc(ncfile)
}
\arguments{
\item{ncfile}{Netcdf file}
}
\value{
Vector of dates
}
\description{
NCtime calculation
}
\author{
Javier Porobic
}
