% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/food.web.R
\name{Tlevel}
\alias{Tlevel}
\title{Throphic level estimation. Equation based on Pauly et al 1998 [Diet composition and trophic levels of marine mammals]}
\usage{
Tlevel(DC, TLp)
}
\arguments{
\item{DC}{Diet composition}

\item{TLp}{Trophic level of the prey}
}
\value{
The trophic level of the predator
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Demiurgo
}

