% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{to.save}
\alias{to.save}
\title{Function to save data}
\usage{
to.save(list, sn = FALSE, rn = FALSE, n = FALSE, b = FALSE,
  Time = Time, tot = FALSE)
}
\arguments{
\item{list}{object from the function}

\item{sn}{Structural weigth Object}

\item{rn}{Reserve weight Object}

\item{n}{Abundance Object}

\item{b}{Biomass Object}

\item{Time}{Time Object}

\item{tot}{Total or partial}
}
\value{
csv file
}
\author{
Demiurgo
}
