% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/food.web.R
\name{tot.prey}
\alias{tot.prey}
\title{Total predation}
\usage{
tot.prey(foc.fg, connect, Fg.code, real.pred, min, grp.dat)
}
\arguments{
\item{foc.fg}{Focus functional group}

\item{connect}{Total connection}

\item{Fg.code}{Original code of the functional groups from the csv file}

\item{real.pred}{Realized predation}

\item{min}{Minimum proportion of the prey in the diet}

\item{grp.dat}{Functional group basic information}
}
\value{
Total consumed prey by predator
}
\description{
Total predation
}
\author{
Demiurgo
}
