% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/food.web.R
\name{trophic.lvl}
\alias{trophic.lvl}
\title{Trophic level}
\usage{
trophic.lvl(rel.prey, grp.dat)
}
\arguments{
\item{rel.prey}{realized diet by predator}

\item{grp.dat}{Functional group basic information}
}
\value{
The trophic level by predator
}
\description{
Trophic level
}
\author{
Javier Porobic
}
