% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catch.R
\name{var.fish}
\alias{var.fish}
\title{Total catch and discard in biomass}
\usage{
var.fish(FG, FISH, nc.data, fsh, grp, is.C = NULL, by.box = FALSE)
}
\arguments{
\item{FG}{Functional groups}

\item{FISH}{Fishery}

\item{nc.data}{NetCDF file with the information of catch}

\item{fsh}{Fisheries csv file (Input from Atlantis)}

\item{grp}{Groups csv file (Input from Atlantis)}

\item{by.box}{Default FALSE. If the information is needed by box (TRUE) or not (FALSE)}

\item{catch}{Default TRUE,  if the data needed is Catch (}

\item{catch}{Default TRUE,  if the data needed is Catch (TRUE) or Discard (FALSE)}
}
\value{
The information of the catch by biomass for all the FG
}
\description{
Total catch and discard in biomass
}
\author{
Demiurgo
}
