% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{bio.age}
\alias{bio.age}
\title{Biomass for age groups}
\usage{
bio.age(age.grp, nc.out, ctg, mg2t, x.cn)
}
\arguments{
\item{age.grp}{Age groups}

\item{nc.out}{ncdf atlantis' output file}

\item{ctg}{category}
}
\value{
a dataframe with the biomass for all the functional groups with age classes
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Demiurgo
}

