% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{bio.pool}
\alias{bio.pool}
\title{Biomass for age groups}
\usage{
bio.pool(pol.grp, nc.out, ctg, mg2t, x.cn, box.info)
}
\arguments{
\item{pol.grp}{Biomass pool groups}

\item{nc.out}{ncdf atlantis' output file}

\item{ctg}{category}

\item{mg2t}{Miligrams to tons scalar}

\item{x.cn}{Ration from nitrogen to carbon}

\item{area}{Box area}

\item{age.grp}{Age groups}
}
\value{
a dataframe with the biomass for all the functional groups with age classes
}
\description{
Biomass for age groups
}
\author{
Demiurgo
}
