% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare}
\alias{compare}
\title{Compare two diferent outputs}
\usage{
compare(nc.out.current, nc.out.old = NULL, grp.csv, bgm.file, cum.depths)
}
\arguments{
\item{nc.out.current}{Current netcdf output file. this is the tradicional output file from Atlantis}

\item{nc.out.old}{(Default = NULL) This is the old netcdf output file,  used only for compare.}

\item{grp.csv}{Groups csv file (Atlatnis input file)}

\item{bgm.file}{XY coordinates bgm file with the information in meters}

\item{cum.depths}{Cummulative depth of the Atlatnis model}
}
\value{
A shiny output (reactive html)
}
\author{
Demiurgo
}
