% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare}
\alias{compare}
\title{Compare Outputs}
\usage{
compare(nc.out.current, nc.out.old = NULL, grp.csv, bgm.file, cum.depths)
}
\arguments{
\item{nc.out.current}{Current netcdf output file. Character string with the path
to the netcdf file to read in. This netcdf file contains is a generic output
from an Atlantis run usually starts with "\emph{output}" and ends in "\emph{.nc}".}

\item{nc.out.old}{(Default = NULL) This is the character string with the path to
the second netcdf output file to read in (used only for comparisons). This
netcdf file contains is a generic output from an Atlantis run and usually
starts with "\emph{output}" and ends in "\emph{.nc}".}

\item{grp.csv}{Character string with the path to the Groups \emph{*.csv}
file (Atlantis input file).}

\item{bgm.file}{Character string with the path to the XY coordinates
Atlantis input file \emph{*.bgm} with the information in meters.}

\item{cum.depths}{Cumulative depth of the Atlantis model}
}
\value{
A reactive HTML with divided in 3 different panels:
\itemize{
\item \bold{Biomass}: This panel shows the results for:
    \itemize{
      \item \emph{Total Biomass} which displays the changes in total biomass for
    each functional group.
      \item \emph{Relative Biomass} displays the variation of the
    total biomass relative to the initial biomass (\eqn{B_{0}}).
     }
 If only one netcdf file is read in it provides a single output, which expands
    two sub-panels in the case of comparisons.
\item \bold{Total}: This function shows the changes in the variables
    (i.e. Biomass, Numbers and Structural and reserve nitrogen) for the entire
    population over time. If a second output is provided the function allows
    comparison of the outputs. In this function the user is able to:
  \itemize{
    \item \emph{visualize any combination of variables}, that means the user can activate and
    deactivate variables from the visualization. This provides more flexibility
    for exploring the changes in the variables and compare between them.
    \item \emph{Scaled}, this option allows for visualization of the values
    normalized by the value in the initial conditions (this option is very
    helpful for checking the main changes in variables during the calibration
    process).
    \item \emph{limit-axis}, indicating whether the user desires total or normalized
    values, this option allows for the y-axis to be unscaled or to set the upper
    limit for the scaled version(from 0 to 3).
   }
\item \bold{By AgeClass}: This tab visualizes the changes through time of the
    variables by age. In this function the user is able to:
    \itemize{
    \item \emph{visualize any combination of variables}, that means that the
    user can activate and deactivate variables from the visualization (this
    provides more flexibility to explore the changes in the variables and compare
    between them).
    \item \emph{Scaled}, this option allows for visualization of the values
    normalised by the value in the initial conditions (this option is very
    helpful for checking the main changes in variables during the calibration
    process).
    \item \emph{limit-axis} which allows the user to set the visualization to
    either total or scaled values (this option allows the y-axis to be unscaled
    or to set the limit from 0 to 3 for the normalized version.
}
}
}
\description{
This function allows visualization of either the output from a simulation or
    comparison between two different Atlantis simulations. The user can compare
    biomasses, abundance, structural or reserve nitrogen between two different
    outputs. Also, the user is able to check the changes is these variables by
    age for the different functional groups.
}
\author{
Demiurgo
}
