% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pprey.mat.R
\name{feeding.mat.shy}
\alias{feeding.mat.shy}
\title{Atlantis feeding tool}
\usage{
feeding.mat.shy(prm.file, grp.file, nc.file, bgm.file, cum.depths,
  quiet = TRUE)
}
\arguments{
\item{prm.file}{Character string with the connection to the Groups \code{*.csv} file (Atlantis input file).}

\item{grp.file}{Character string with the connection to the Groups \code{*.csv} file (Atlantis input file).}

\item{nc.file}{Character string with the connection to the netcdf file to read in. This netcdf file contains the initial conditions for the Atlantis model usually ends in \code{.nc}".}

\item{bgm.file}{Character string with the connection to the XY coordinates Antlantis input file \code{*.bgm} with the information in meters.}

\item{cum.depths}{Vector with the cumulative depths of the differente layers \code{cum.depths <- c(0, 20, 100, 200, 500)}}

\item{quiet}{(Default = TRUE) this paramter helps during the process of debuging.}
}
\value{
Reactive html that display predator prey information such as: the ppPREY matrix,  the initial abundace of prey, the overlap matrix based on gape size, predator preference and predator prey spatial overlap.
}
\author{
Demiurgo
}
