% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pprey.mat.R
\name{feeding.mat.shy}
\alias{feeding.mat.shy}
\title{Atlantis feeding tool}
\usage{
feeding.mat.shy(prm.file, grp.file, nc.file, bgm.file, cum.depths,
  quiet = TRUE)
}
\arguments{
\item{prm.file}{Atlantis parameter file}

\item{grp.file}{Atlantis group file}

\item{nc.file}{Atlantis initial conditions file}

\item{bgm.file}{Bgm used for the Atlantis model}

\item{cum.depths}{Cumulative depths of the differente layers}

\item{quiet}{For debugging porpouses,  default TRUE}
}
\value{
Reactive html that display predator prey information such as: the ppPREY matrix,  the initial abundace of prey, the overlap matrix based on gape size, predator preference and predator prey spatial overlap.
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Demiurgo
}

