% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/food.web.R
\name{food.web}
\alias{food.web}
\title{Trophic level of the fucntional groups}
\usage{
food.web(diet.file, grp.file, quiet = TRUE)
}
\arguments{
\item{diet.file}{Diet output file from atlantis}

\item{grp.file}{Character string with the connection to the Groups \code{*.csv} file (Atlantis input file).}

\item{quiet}{(Default = TRUE) this paramter helps during the process of debuging.}
}
\value{
Reactive output
}
\author{
Demiurgo
}
