% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/food.web.R
\name{food.web}
\alias{food.web}
\title{Function on shiny R to plot the trophic level of the fucntional groups}
\usage{
food.web(diet.file, grp.file, quiet = TRUE)
}
\arguments{
\item{diet.file}{Diet output file from atlantis}

\item{grp.file}{Group file. Atlantis input file (.cvs)}

\item{quiet}{help to debug the function}
}
\value{
Reactive output
}
\author{
Demiurgo
}
