% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth.R
\name{growth.pp}
\alias{growth.pp}
\title{Growth primary producers}
\usage{
growth.pp(ini.nc.file, grp.file, prm.file, out.nc.file)
}
\arguments{
\item{ini.nc.file}{Character string with the connection to the netcdf file to read in. This netcdf file contains the initial conditions for the Atlantis model usually ends in \code{.nc}".}

\item{grp.file}{Character string with the connection to the Groups \code{*.csv} file (Atlantis input file).}

\item{prm.file}{Character string with the connection to the biology parameter file \code{*.prm}.}

\item{out.nc.file}{Character string with the connection to the netcdf file to read in. This netcdf file contains is a generic output from an Atlantis run usually starts with \code{output} and ends in \code{.nc}".}
}
\value{
A shiny output (reactive html)
}
\author{
Demiurgo
}
