% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth.R
\name{growth.pp}
\alias{growth.pp}
\title{Limitation Growth primary producers}
\usage{
growth.pp(ini.nc.file, grp.file, prm.file, out.nc.file)
}
\arguments{
\item{ini.nc.file}{Character string with the path to the netcdf file to
read in. This netcdf file contains the initial conditions for the Atlantis
model usually ends in \emph{.nc}.}

\item{grp.file}{Character string with the path to the Groups \emph{*.csv}
file (Atlantis input file).}

\item{prm.file}{Character string with the path to the biology parameter
file \emph{*.prm}.}

\item{out.nc.file}{Character string with the path to the netcdf file to
read in. This netcdf file contains is a generic output from an Atlantis run
usually starts with \emph{output} and ends in \emph{.nc}.}
}
\value{
A reactive HTML with graphical output by functional group, layer and box
    of the variables:
\itemize{
  \item \bold{Growth}:  This variable is the  Primary producer growth (\eqn{G_{pp}}) which is
    determined by the biomass for the primary producer (\eqn{Biom_{pp}}),  the maximum effective
    growth (\eqn{mum}) and the limiting factors: Light limitation (\eqn{\delta_{light}}),
    Nutrients limitation \eqn{\delta_{nutrients}}, and Eddy scalar (\eqn{\delta_{eddy}}).
    \deqn{G_{pp} = Biom_{pp} * mum * \delta_{light} * \delta_{nutrients} *
    \delta_{eddy}}
\item \bold{Light limitation}: The available light by box and layer this option
    use only the basic formula and currently does not use the light adaptation
    option. The light attenuation values used for the calculations are the values
    estimated by Atlantis and reported in the \bold{\emph{out.nc.file}}.
\item \bold{Nutrients limitation}:  Nutrients limitation calculated by layer and
    by box, using the option chosen by the user in the Atlantis run
    (\emph{flagnut = 0,  1 or 2}).
\item \bold{Eddy scalar}: This value is calculated by functional group and by
    box. The values of eddies strength (\eqn{eddy_{strength}}) used are from the
    Atlantis output and the value of eddy scale (\eqn{eddy_{scale}}) is read from
    the Atlantis configuration file.
  }
}
\description{
This function helps you visualize what are the possible factors that limit the
    growth of primary producers. Primary producers in Atlantis are limited in
    their biomass by variables such as: nutrients, light, space, and by processes
    such as acidification and the presence of eddies (Audijyontze et al. 2017). In this
    function, most of these factors are analyzed, except for the acidification
    process and the spatial limitation.
\itemize{
   \item \bold{Nutrient limitation}: The nutrients limitation scalar
    (\eqn{\delta_{nutrients}}) in the case of species limited only by nitrogen is
    calculated as:
    \deqn{\delta_{nutrients}  = \frac{DIN}{KN + DIN}} \cr
    were \eqn{DIN} is the concentration of \eqn{NH_{3}} + \eqn{NO_{3}} and
    \eqn{KN} is the half-saturation constant of nutrient uptake. \cr
    For species that are limited for multiple nutrients like diatom that can be
    limited by silicate \eqn{Si} and nitrogen you have 3 different options to choose on
    Atlantis (\emph{flagnut}) and the nutrient limitation can have different options:
    \itemize{
      \item \emph{Leibig limitation \bold{flagnut = 0}}
      \deqn{\delta_{nutrients} = min(\frac{DIN}{KN + DIN}, \frac{Si}{KS + Si})}
      \item \emph{multiplicative limitation \bold{flagnut = 1}}
      \deqn{\delta_{nutrients} = \sqrt{\frac{DIN}{KN + DIN} * \frac{Si}{KS + Si}}}
      \item \emph{ERSEM WQI limitation \bold{flagnut = 2}}
      \deqn{\delta_{nutrients} = \frac{2}{\frac{DIN}{KN + DIN} * \frac{Si}{KS + Si}}}
    were \eqn{KS} is the growth half-saturation constant for the functional group.
           }
  \item \bold{Light limitation}: In Atlantis photosynthesis in  primary producers
    can be limited by light. The light limitation factor \eqn{\delta_{light}} is calculated by:
 \deqn{\delta_{light} = min(\frac{IRR}{KI}, 1)} \cr
 were \eqn{IRR} irradiance or available light  and\eqn{KI} is the light
    saturation coefficient. \cr
In Atlantis there is a second option \emph{flaglight = 1} that allows the calculation of the light
    limitation factor allowing for light adaptation. This is intended to capture
    the ability of the primary producers to rapidly adapt to different light
    condition (Audijyontze et al. 2017). This option was not considered in this version
    of the tool but would be implemented in future updates.
 \item \bold{Effect of eddies on primary production}: The eddy effect on primary
    produces \eqn{\delta_{eddy}}  is calculated by multiplying the scale parameter \eqn{eddy_{scale}}
    (\emph{eddy_scale} on Atlantis) by the eddy strength \eqn{eddy_{strength}}.
   \deqn{\delta_{eddy} = eddy_{scale} * eddy_{strength} }
}
}
\author{
Demiurgo
}
