% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predation.R
\name{predation}
\alias{predation}
\title{Predation}
\usage{
predation(biom.file, groups.csv, diet.file, age.biomass = NULL)
}
\arguments{
\item{biom.file}{Output biomass file from the Atlantis simulation. Usually this output has the name 'outputYOUR_Model_NAMEBiomIndx.txt',  were YOUR_Model_NAME is the name of your Atlatnis model.}

\item{groups.csv}{This file is the group csv file. Which is one of the compulsory inputs of Atlantis}

\item{diet.file}{Usual Atlatnis output file,. This file contain the diets of each functional group at each (recorded) time step. If the atlantis simulation is for too many years, it is highly recommended a low frequency recording periodicity of the output file (toutinc). General high frequency engravings very large files and difficult to handle in R.}

\item{age.biomass}{output file for the biomass by age (non-standard atlantis output, make sure to put the flag 'flag_age_output' in the run.prm if you want to look at this output)}
}
\value{
A reactive html output useful for dynamically evaluating and visualizing predator-prey relationships from atlantis output.
}
\author{
Demiurgo
}

