% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/food.web.R
\name{prey.pos}
\alias{prey.pos}
\title{Assing trophic level of the prey the values are based on Pauly et al 1998 [Diet composition and trophic levels of marine mammals]
       And Tucket and Rogers 2014 [Examining predator–prey body size, trophic level and body mass across marine and terrestrial mammals]}
\usage{
prey.pos(FGs, grp.dat)
}
\arguments{
\item{FGs}{Preys}

\item{grp.dat}{csv groups file}
}
\value{
The trophic level of the Functional group
}
\description{
Assing trophic level of the prey the values are based on Pauly et al 1998 [Diet composition and trophic levels of marine mammals]
       And Tucket and Rogers 2014 [Examining predator–prey body size, trophic level and body mass across marine and terrestrial mammals]
}
\author{
Demiurgo
}
