% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{relative}
\alias{relative}
\title{Calculate the value of a time serie based on the first value}
\usage{
relative(df, biomass = TRUE, Vec = NULL)
}
\arguments{
\item{df}{Data frame of biomass by FG}

\item{biomass}{Is biomass included}

\item{Vec}{True is the information in vector}
}
\value{
A vector of relative values
}
\description{
Calculate the value of a time serie based on the first value
}
\author{
Demiurgo
}
