% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R, R/growth.R, R/pprey.mat.R,
%   R/recruitment.R
\name{text2num}
\alias{text2num}
\alias{text2num}
\alias{text2num}
\alias{text2num}
\title{Parameter file reader}
\usage{
text2num(text, pattern, FG = NULL, Vector = FALSE)

text2num(text, pattern, FG = NULL, Vector = FALSE)

text2num(text, pattern, FG = NULL, Vector = FALSE)

text2num(text, pattern, FG = NULL, Vector = FALSE)
}
\arguments{
\item{text}{Biological parametar file for Atlatnis}

\item{pattern}{Text that you are looking}

\item{FG}{Name of the functional groups}

\item{Vector}{Logic argument, if the data is on vectors or not}

\item{text}{Biological parametar file for Atlantis}

\item{pattern}{Text that you are looking}

\item{FG}{Name of the functional groups}

\item{Vector}{Logic argument, if the data is on vectors or not}

\item{text}{Biological parametar file for Atlatnis}

\item{pattern}{Text that you are looking}

\item{FG}{Name of the functional groups}

\item{Vector}{Logic argument, if the data is on vectors or not}

\item{text}{Biological parametar file for Atlatnis}

\item{pattern}{Text that you are looking}

\item{FG}{Name of the functional groups}

\item{Vector}{Logic argument, if the data is on vectors or not}
}
\value{
A matrix with the values from the .prm file

A matrix with the values from the .prm file

A matrix with the values from the .prm file

A matrix with the values from the .prm file
}
\author{
Demiurgo

Demiurgo

Demiurgo

Demiurgo
}
